+++ src/Animation/Variables.js
@@ -16,8 +16,11 @@ const slideLeft = {
         }
     },
     exit: {
-        x: "100vw",
-        transition: { ease: "easeInOut" }
+        x: "-100vw",
+        transition: {
+            ease: "easeInOut",
+            duration: 1
+        }
     }
 }
 
@@ -40,7 +43,36 @@ const slideRight = {
     },
     exit: {
         x: "-100vw",
-        transition: { ease: "easeInOut" }
+        transition: {
+            ease: "easeInOut",
+            duration: 1
+        }
+    }
+}
+
+const slideBottom = {
+    init: {
+        opacity: 0,
+        y: "-100vw"
+    },
+    visible: {
+        opacity: 1,
+        y: 0,
+        transition: {
+            type: "spring",
+            mass: 0.8,
+            damping: 10,
+            when: "beforeChildren",
+            delay: 0,
+            duration: 1
+        }
+    },
+    exit: {
+        y: "-100vw",
+        transition: {
+            ease: "easeInOut",
+            duration: 1
+        }
     }
 }
 
@@ -62,8 +94,28 @@ const opacity = {
     }
 }
 
+const opacityHome = {
+    init: {
+        opacity: 0,
+    },
+    visible: {
+        opacity: 1,
+        transition: { delay: 1, duration: 1 }
+    },
+    exit: {
+        x: "-100vw",
+        transition: {
+            duration: 0.7,
+            ease: "easeInOut",
+            when: "beforeChildren",
+        }
+    }
+}
+
 module.exports = {
     slideLeft,
     slideRight,
-    opacity
+    slideBottom,
+    opacity,
+    opacityHome
 }